+++ src/libopenrave/robotmanipulator.cpp
@@ -587,6 +587,48 @@ bool RobotBase::Manipulator::CheckEndEffectorCollision(const Transform& tEE, Col
     return false;
 }
 
+bool RobotBase::Manipulator::CheckEndEffectorSelfCollision(const Transform& tEE, CollisionReportPtr report) const
+{
+    RobotBasePtr probot(__probot);
+    Transform toldEE = GetTransform();
+    Transform tdelta = tEE*toldEE.inverse();
+    // get all child links of the manipualtor
+    int iattlink = __pEffector->GetIndex();
+    vector<LinkPtr> vattachedlinks;
+    __pEffector->GetRigidlyAttachedLinks(vattachedlinks);
+    FOREACHC(itlink,vattachedlinks) {
+        if( probot->CheckLinkSelfCollision((*itlink)->GetIndex(),tdelta*(*itlink)->GetTransform(),report) ) {
+            return true;
+        }
+    }
+    FOREACHC(itlink, probot->GetLinks()) {
+        int ilink = (*itlink)->GetIndex();
+        if((ilink == iattlink)|| !(*itlink)->IsEnabled() ) {
+            continue;
+        }
+        // gripper needs to be affected by all joints
+        bool bGripperLink = true;
+        FOREACHC(itarmdof,__varmdofindices) {
+            if( !probot->DoesAffect(probot->GetJointFromDOFIndex(*itarmdof)->GetJointIndex(),ilink) ) {
+                bGripperLink = false;
+                break;
+            }
+        }
+        if( !bGripperLink ) {
+            continue;
+        }
+        for(size_t ijoint = 0; ijoint < probot->GetJoints().size(); ++ijoint) {
+            if( probot->DoesAffect(ijoint,ilink) && !probot->DoesAffect(ijoint,iattlink) ) {
+                if( probot->CheckLinkSelfCollision(ilink,tdelta*(*itlink)->GetTransform(),report) ) {
+                    return true;
+                }
+                break;
+            }
+        }
+    }
+    return false;
+}
+
 bool RobotBase::Manipulator::CheckEndEffectorCollision(const IkParameterization& ikparam, CollisionReportPtr report) const
 {
     if( ikparam.GetType() == IKP_Transform6D ) {
@@ -629,6 +671,57 @@ bool RobotBase::Manipulator::CheckEndEffectorCollision(const IkParameterization&
         }
         if( !bhassimilar ) {
             if( CheckEndEffectorCollision(GetTransform(),report) ) {
+                return true;
+            }
+            listprevtransforms.push_back(t);
+        }
+    }
+
+    return false;
+}
+
+bool RobotBase::Manipulator::CheckEndEffectorSelfCollision(const IkParameterization& ikparam, CollisionReportPtr report) const
+{
+    if( ikparam.GetType() == IKP_Transform6D ) {
+        return CheckEndEffectorSelfCollision(ikparam.GetTransform6D(),report);
+    }
+    RobotBasePtr probot = GetRobot();
+    IkSolverBasePtr pIkSolver = GetIkSolver();
+    OPENRAVE_ASSERT_OP_FORMAT((int)GetArmIndices().size(), <=, ikparam.GetDOF(), "ikparam type 0x%x does not fully determine manipulator %s:%s end effector configuration", ikparam.GetType()%probot->GetName()%GetName(),ORE_InvalidArguments);
+    OPENRAVE_ASSERT_FORMAT(!!pIkSolver, "manipulator %s:%s does not have an IK solver set",probot->GetName()%GetName(),ORE_Failed);
+    OPENRAVE_ASSERT_FORMAT(pIkSolver->Supports(ikparam.GetType()),"manipulator %s:%s ik solver %s does not support ik type 0x%x",probot->GetName()%GetName()%pIkSolver->GetXMLId()%ikparam.GetType(),ORE_InvalidState);
+    BOOST_ASSERT(pIkSolver->GetManipulator() == shared_from_this() );
+    IkParameterization localgoal;
+    if( !!__pBase ) {
+        localgoal = __pBase->GetTransform().inverse()*ikparam;
+    }
+    else {
+        localgoal=ikparam;
+    }
+
+    // only care about the end effector position, so disable all time consuming options. still leave the custom options in case the user wants to call some custom stuff?
+    // is it necessary to call with IKFO_IgnoreJointLimits knowing that the robot will never reach those solutions?
+    std::vector< std::vector<dReal> > vsolutions;
+    if( !pIkSolver->SolveAll(localgoal, vector<dReal>(), IKFO_IgnoreSelfCollisions,vsolutions) ) {
+        throw OPENRAVE_EXCEPTION_FORMAT("failed to find ik solution for type 0x%x",ikparam.GetType(),ORE_InvalidArguments);
+    }
+    RobotStateSaver saver(probot);
+    probot->SetActiveDOFs(GetArmIndices());
+    // have to check all solutions since the 6D transform can change even though the ik parameterization doesn't
+    std::list<Transform> listprevtransforms;
+    FOREACH(itsolution,vsolutions) {
+        probot->SetActiveDOFValues(*itsolution,false);
+        Transform t = GetTransform();
+        // check if previous transforms exist
+        bool bhassimilar = false;
+        FOREACH(ittrans,listprevtransforms) {
+            if( TransformDistanceFast(t,*ittrans) < g_fEpsilonLinear*10 ) {
+                bhassimilar = true;
+                break;
+            }
+        }
+        if( !bhassimilar ) {
+            if( CheckEndEffectorSelfCollision(GetTransform(),report) ) {
                 return true;
             }
             listprevtransforms.push_back(t);